channels: extract github permission-guidance into its own module
parseListHooksPermissionStatus and buildPermissionGuidance lived in
src/channels/adapters/github/index.ts even though their test
(permission-guidance.test.ts) and consumer (logRegistrationOutcome) are
the only callers. Two upcoming changes need to call the same helpers:

  1. A startup permission preflight that reads the installation's
     granted-permission map and warns when a configured event in
     typeclaw.json#channels.github.eventAllowlist would never deliver.
  2. Outbound 403 guidance that translates 'Resource not accessible by
     integration' into a per-endpoint 'enable X permission' instruction.

Both want the same auth-type-aware (pat vs app) formatting and the same
verbatim github.com UI labels. Move the helpers into
permission-guidance.ts so neither new caller has to import from index.ts
(which would pull in the whole adapter factory) and so the upcoming
endpoint-to-permission table can live next to the existing
status-to-guidance text. Pure refactor — no behavior change.
